apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.19
pydantic-settings==2.3.1
anthropic==0.28.0
httpx==0.27.0
feedparser==6.0.11
python-frontmatter==1.1.0
apscheduler==3.10.4
watchdog==4.0.1
python-multipart==0.0.9
# Arize AX observability
arize-otel
openinference-instrumentation-anthropic
opentelemetry-sdk
beautifulsoup4
lxml
aiofiles
python-dotenv
# Document parsing for content uploads
pypdf>=4.0.0
python-docx>=1.1.0
# Google Calendar integration
google-auth-oauthlib>=1.2.0
# Token encryption for OAuth credentials
cryptography>=42.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
pytrends>=4.7.0
pytest>=8.0.0
PyJWT>=2.8.0
stripe>=10.0.0
